distsqlrun: split out scrub tablereader
Previously, tablereader had logic for both scrub and normal operation
inside of it. This was hard to read and probably slightly inefficient
due to the extra switching that happened per row.

Now, the scrub table reader lives in a separate file and is instantiated
at plan time.

Release note: None
